#902 WIP: Address win32/x86 build warnings

+++ src/netxs/desktopio/geometry.hpp
@@ -831,7 +831,7 @@ namespace netxs
             auto h = area.size.y;
             auto image_raster = raster<irgb>();
             auto data = image_raster._data;
-            assert(area.length() <= data.size()); // Invalid sprite element type.
+            assert((size_t)area.length() <= data.size()); // Invalid sprite element type.
             auto get_row = [&](si32 y){ return data.subspan(y * w, w); };
             auto y0 = 0;
             auto is_visible = [](irgb p){ return p.a != 0; };
